Refactor shadow location imports to use centralized SHADOW_LOCATION_PREFIX constant
- Updated multiple components and hooks to import SHADOW_LOCATION_PREFIX from '@/app/lib/shadowConstants' for consistency.
- Removed hardcoded '🔮' prefix references to enhance maintainability and ensure a single source of truth for shadow location naming.
