+++ src/app/lib/shadowConstants.ts
@@ -1,0 +1,1 @@
+export const SHADOW_LOCATION_PREFIX = '🔮';
